:sparkles: Add macOS support
