+++ .github/workflows/test.yml
@@ -1,9 +1,18 @@
-name: Test Action
+name: Test
 
 on: push
 
 jobs:
-  test:
+  macOS:
+    runs-on: macos-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: ./
+        with:
+          cljstyle-version: 0.13.0
+      - run: cljstyle version
+
+  linux:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
