+++ package.json
@@ -10,7 +10,7 @@
     "lint": "eslint './**/*.ts'",
     "format": "eslint --fix './**/*.ts'",
     "typecheck": "tsc",
-    "build": "esbuild src/setup-cljstyle.ts --bundle --outdir=dist --target=node12 --platform=node",
+    "build": "esbuild src/index.ts --outfile=dist/setup-cljstyle.js --minify --bundle --target=node12 --platform=node",
     "git-add": "git add dist/setup-cljstyle.js"
   },
   "pre-commit": [
