+++ src/installers/index.ts
@@ -1,0 +1,3 @@
+export { setupOnWindows } from './windows';
+export { setupOnMacOS } from './macos';
+export { setupOnLinux } from './linux';
