+++ src/installers/windows.ts
@@ -1,0 +1,3 @@
+export const setupOnWindows = async (): Promise<void> => {
+    // TODO: Add Windows support
+};
