+++ tsconfig.json
@@ -1,45 +1,16 @@
 {
   "compilerOptions": {
-    /* 基本オプション */
-    "target": "es5",                           /* ECMAScript のターゲット バージョンを指定します: 'ES3' (既定値)、'ES5'、'ES2015'、'ES2016'、'ES2017'、'ES2018'、'ES2019'、'ES2020'、'ESNEXT'。 */
-    "module": "commonjs",                      /* モジュール コードの生成を指定します: 'none'、'commonjs'、'amd'、'system'、'umd'、'es2015'、'es2020、'ESNext'。 */
-    "noEmit": true,                         /* 出力しないでください。 */
-
-    /* 詳細オプション */
-    "locale": "ja",                            /* ユーザーにメッセージを表示するときに使用するロケール (例: 'en-us') */
-    "skipLibCheck": true,                      /* 宣言ファイルの型チェックをスキップします。 */
-    "forceConsistentCasingInFileNames": true,  /* 同じファイルへの大文字小文字の異なる参照を許可しない。 */
-
-    /* Strict 型チェック オプション */
-    "strict": true,                            /* 厳密な型チェックのオプションをすべて有効にします。 */
-
-    /* 追加のチェック */
-    "noUnusedLocals": true,                 /* 使用されていないローカルに関するエラーを報告します。 */
-    "noUnusedParameters": true,             /* 使用されていないパラメーターに関するエラーを報告します。 */
-    "noImplicitReturns": true,              /* 関数の一部のコード パスが値を返さない場合にエラーを報告します。 */
-    "noFallthroughCasesInSwitch": true,     /* switch ステートメントに case のフォールスルーがある場合にエラーを報告します。 */
-
-    /* モジュール解決のオプション */
-    // "moduleResolution": "node",             /* モジュールの解決方法を指定します: 'node' (Node.js) または 'classic' (TypeScript pre-1.6)。 */
-    // "baseUrl": "./",                        /* 相対モジュール名を解決するためのベース ディレクトリ。 */
-    // "paths": {},                            /* 'baseUrl' の相対的な場所を検索するためにインポートを再マップする一連のエントリ。 */
-    // "rootDirs": [],                         /* 結合されたコンテンツがランタイムでのプロジェクトの構成を表すルート フォルダーの一覧。 */
-    // "typeRoots": [],                        /* 含める型定義の元のフォルダーの一覧。 */
-    // "types": [],                            /* コンパイルに含む型宣言ファイル。 */
-    // "allowSyntheticDefaultImports": true,   /* 既定のエクスポートがないモジュールからの既定のインポートを許可します。これは、型チェックのみのため、コード生成には影響を与えません。 */
-    "esModuleInterop": true                    /* すべてのインポートの名前空間オブジェクトを作成して、CommonJS と ES モジュール間の生成の相互運用性を有効にします。'allowSyntheticDefaultImports' を暗黙のうちに表します。 */
-    // "preserveSymlinks": true,               /* symlink の実際のパスを解決しません。 */
-    // "allowUmdGlobalAccess": true,           /* モジュールから UMD グローバルへのアクセスを許可します。 */
-
-    /* ソース マップ オプション */
-    // "sourceRoot": "",                       /* デバッガーがソースの場所の代わりに TypeScript ファイルを検索する必要のある場所を指定します。 */
-    // "mapRoot": "",                          /* デバッガーが、生成された場所の代わりにマップ ファイルを検索する必要のある場所を指定します。 */
-    // "inlineSourceMap": true,                /* 個々のファイルを持つ代わりに、複数のソース マップを含む単一ファイルを生成します。 */
-    // "inlineSources": true,                  /* 単一ファイル内でソースマップと共にソースを生成します。'--inlineSourceMap' または '--sourceMap' を設定する必要があります。 */
-
-    /* 試験的なオプション */
-    // "experimentalDecorators": true,         /* ES7 デコレーター用の実験的なサポートを有効にします。 */
-    // "emitDecoratorMetadata": true,          /* デコレーター用の型メタデータを発行するための実験的なサポートを有効にします。 */
+    "target": "es5",
+    "module": "commonjs",
+    "noEmit": true,
+    "locale": "ja",
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noImplicitReturns": true,
+    "noFallthroughCasesInSwitch": true,
   },
   "include": [
     "src"
